# docs/swagger.yaml
basePath: /
definitions:
...
  models.Menu:
    properties:
      category:
        type: string
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      menu_id:
        type: string
      name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
    required:
    - category
    - name
    type: object
  models.Order:
    properties:
      created_at:
        type: string
      id:
        type: string
      order_date:
        type: string
      order_id:
        type: string
      table_id:
        type: string
      updated_at:
        type: string
    required:
    - order_date
    - table_id
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: CRUP api's for go backend
  title: go-backend
  version: "1.0"
paths:
...
  /menu/{menu_id}:
    patch:
      consumes:
      - application/json
      description: Updates an existing menu resource in the menu collection
      parameters:
      - description: ID of the menu resource to update
        in: path
        name: menu_id
        required: true
        type: string
      - description: Menu object
        in: body
        name: Menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: menu got updated with new body
          schema:
            $ref: '#/definitions/models.Menu'
        "404":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error while updating food
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Updates an menu resource
      tags:
      - menu
  /menus:
    get:
      description: Returns an array of menus from the menucollection in restaurent
        database.
      produces:
      - application/json
      responses:
        "200":
          description: 'slice of menus '
          schema:
            items:
              $ref: '#/definitions/models.Menu'
            type: array
      security:
      - Security.require(true): []
      summary: Returns slice of menus
      tags:
      - menu
    post:
      consumes:
      - application/json
      description: Creates a menu resource on the server
      parameters:
      - description: Menu object
        in: body
        name: Menu
        required: true
        schema:
          $ref: '#/definitions/models.Menu'
      produces:
      - application/json
      responses:
        "201":
          description: New menu created
          schema:
            $ref: '#/definitions/models.Menu'
        "500":
          description: Internal Server Error while creating a new menu
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Creates a new menu
      tags:
      - menu
  /menus/{menu_id}:
    get:
      consumes:
      - application/json
      description: Retrieves a menu with specific menu id from the menu collection
      parameters:
      - description: menu_id
        in: path
        name: menu_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Details of a specific menu
          schema:
            $ref: '#/definitions/models.Menu'
        "500":
          description: Internal Server Error in mongodb
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Retrieves a menu  with specific menu id
      tags:
      - menu
...
  /orders:
    get:
      description: Returns an array of orders placed from the ordercollection in restaurent
        database.
      produces:
      - application/json
      responses:
        "200":
          description: 'slice of orders '
          schema:
            items:
              $ref: '#/definitions/models.Order'
            type: array
      security:
      - Security.require(true): []
      summary: Returns slice of orders
      tags:
      - order
    post:
      consumes:
      - application/json
      description: Creates a order resource on the server
      parameters:
      - description: Order object
        in: body
        name: Order
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "201":
          description: New order created
          schema:
            $ref: '#/definitions/models.Order'
        "500":
          description: Internal Server Error while creating a new order
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Creates a order  resource
      tags:
      - order
  /orders/{order_id}:
    get:
      consumes:
      - application/json
      description: Retrieves a order with specific order id from the orders collection
      parameters:
      - description: order_id
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Details of a specific order
          schema:
            $ref: '#/definitions/models.Order'
        "500":
          description: Internal Server Error in mongodb
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Retrieves a order with specific order id
      tags:
      - order
    patch:
      consumes:
      - application/json
      description: Updates an existing order resource in the orders collection
      parameters:
      - description: ID of the order resource to update
        in: path
        name: order_id
        required: true
        type: string
      - description: Order object
        in: body
        name: Order
        required: true
        schema:
          $ref: '#/definitions/models.Order'
      produces:
      - application/json
      responses:
        "200":
          description: orders got updated with new body
          schema:
            $ref: '#/definitions/models.Order'
        "404":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error while updating order
          schema:
            type: string
      security:
      - Security.require(true): []
      summary: Updates an order  resource
      tags:
      - order
...
securityDefinitions:
  "":
    description: JWT token for authentication
    in: header
    name: token
    type: apiKey
swagger: "2.0"